feat: json serializer for complex fields
This is a breaking change. Models stored in redis with previous versions
of pydantic-aioredis won't properly deserialize with this change in
place.

The new serializer uses json to dump to string any complex objects
(List, Dict, Set, Python object, pydantic models, etc) and then stores
this string as the field value in the hash in redis. On retrieval from
redis, this string is json.loaded.